Add vector aggregation tests

+++ src/main/java/io/atoti/spark/Discovery.java
@@ -30,7 +30,7 @@ public class Discovery {
                 DataTypes.ShortType,
                 DataTypes.StringType,
                 DataTypes.TimestampType,
-                DataTypes.createArrayType(DataTypes.IntegerType),
+                DataTypes.createArrayType(DataTypes.LongType),
               })
           .collect(Collectors.toMap(DataType::toString, Function.identity()));
